requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
psycopg==3.1.19
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.2
